From: Johannes Pelto-Piri <johannes.peltopiri@gmail.com> Date: Fri, 11 May 2012 13:32:14 +0200 Subject: [PATCH] Fixed so that a multi line header works.
This is done by redefining the frametitle command so that the frame title
is printed using two \parbox commands. This is a pretty ugly solution but
it works until we have three lines in the heading.

The solution is this is based on [1] also used section numbers, which could
be added pretty easily if needed.

[1] http://tex.stackexchange.com/questions/47411/how-to-indent-long-frametitle-beamer